+++ src/dftbp/dftbplus/parser.F90
@@ -8407,10 +8407,21 @@ contains
             call convertUnitHsd(char(modifier), timeUnits, child3, inp%coupling)
             inp%coupling = deltaT / inp%coupling
           else
-            call error("Either CouplingStrength or Timescale must be set for Berendsen thermostats.")
+            call error("Either CouplingStrength or Timescale must be set for Berendsen&
+                & thermostats.")
           end if
         end if
       end associate
+
+    case ("langevin")
+
+      thermostatInp%thermostatType = thermostatTypes%langevin
+      allocate(thermostatInp%langevin)
+      call readTempOrTempProfile_(thermNode, maxRun, tempProfileInp)
+      call getChildValue(thermNode, "timeconstant", thermostatInp%langevin%gamma,&
+          & modifier=modifier, child=child3)
+      call convertUnitHsd(char(modifier), timeUnits, child3, thermostatInp%langevin%gamma)
+      thermostatInp%langevin%gamma = 1.0_dp / thermostatInp%langevin%gamma
 
     case ("nosehoover")